CM-1631: Fixed Experiment to return correct Gig meta-data object from public API. Fixed related test cases. Fixed CometApi to return ExperimentMetadata from public API. Fixed related test cases. Renamed internal model package to the rest for better clarity.
