+++ comet-java-client/src/main/java/ml/comet/experiment/impl/BaseExperimentAsync.java
@@ -10,12 +10,12 @@ import lombok.NonNull;
 import ml.comet.experiment.context.ExperimentContext;
 import ml.comet.experiment.impl.asset.Asset;
 import ml.comet.experiment.impl.asset.RemoteAsset;
-import ml.comet.experiment.impl.model.HtmlRest;
-import ml.comet.experiment.impl.model.LogDataResponse;
-import ml.comet.experiment.impl.model.LogOtherRest;
-import ml.comet.experiment.impl.model.MetricRest;
-import ml.comet.experiment.impl.model.OutputUpdate;
-import ml.comet.experiment.impl.model.ParameterRest;
+import ml.comet.experiment.impl.rest.HtmlRest;
+import ml.comet.experiment.impl.rest.LogDataResponse;
+import ml.comet.experiment.impl.rest.LogOtherRest;
+import ml.comet.experiment.impl.rest.MetricRest;
+import ml.comet.experiment.impl.rest.OutputUpdate;
+import ml.comet.experiment.impl.rest.ParameterRest;
 import ml.comet.experiment.impl.utils.AssetUtils;
 import ml.comet.experiment.model.GitMetaData;
 import org.slf4j.Logger;
