+++ comet-java-client/src/main/java/ml/comet/experiment/impl/RestApiClient.java
@@ -10,29 +10,29 @@ import ml.comet.experiment.impl.asset.RemoteAsset;
 import ml.comet.experiment.impl.constants.FormParamName;
 import ml.comet.experiment.impl.constants.QueryParamName;
 import ml.comet.experiment.impl.http.Connection;
-import ml.comet.experiment.impl.model.AddExperimentTagsRest;
-import ml.comet.experiment.impl.model.AddGraphRest;
-import ml.comet.experiment.impl.model.CreateExperimentRequest;
-import ml.comet.experiment.impl.model.CreateExperimentResponse;
-import ml.comet.experiment.impl.model.ExperimentAssetListResponse;
-import ml.comet.experiment.impl.model.ExperimentMetadataRest;
-import ml.comet.experiment.impl.model.ExperimentStatusResponse;
-import ml.comet.experiment.impl.model.ExperimentTimeRequest;
-import ml.comet.experiment.impl.model.GetExperimentsResponse;
-import ml.comet.experiment.impl.model.GetGraphResponse;
-import ml.comet.experiment.impl.model.GetHtmlResponse;
-import ml.comet.experiment.impl.model.GetOutputResponse;
-import ml.comet.experiment.impl.model.GetProjectsResponse;
-import ml.comet.experiment.impl.model.GetWorkspacesResponse;
-import ml.comet.experiment.impl.model.GitMetadataRest;
-import ml.comet.experiment.impl.model.HtmlRest;
-import ml.comet.experiment.impl.model.LogDataResponse;
-import ml.comet.experiment.impl.model.LogOtherRest;
-import ml.comet.experiment.impl.model.MetricRest;
-import ml.comet.experiment.impl.model.MinMaxResponse;
-import ml.comet.experiment.impl.model.OutputUpdate;
-import ml.comet.experiment.impl.model.ParameterRest;
-import ml.comet.experiment.impl.model.TagsResponse;
+import ml.comet.experiment.impl.rest.AddExperimentTagsRest;
+import ml.comet.experiment.impl.rest.AddGraphRest;
+import ml.comet.experiment.impl.rest.CreateExperimentRequest;
+import ml.comet.experiment.impl.rest.CreateExperimentResponse;
+import ml.comet.experiment.impl.rest.ExperimentAssetListResponse;
+import ml.comet.experiment.impl.rest.ExperimentMetadataRest;
+import ml.comet.experiment.impl.rest.ExperimentStatusResponse;
+import ml.comet.experiment.impl.rest.ExperimentTimeRequest;
+import ml.comet.experiment.impl.rest.GetExperimentsResponse;
+import ml.comet.experiment.impl.rest.GetGraphResponse;
+import ml.comet.experiment.impl.rest.GetHtmlResponse;
+import ml.comet.experiment.impl.rest.GetOutputResponse;
+import ml.comet.experiment.impl.rest.GetProjectsResponse;
+import ml.comet.experiment.impl.rest.GetWorkspacesResponse;
+import ml.comet.experiment.impl.rest.GitMetadataRest;
+import ml.comet.experiment.impl.rest.HtmlRest;
+import ml.comet.experiment.impl.rest.LogDataResponse;
+import ml.comet.experiment.impl.rest.LogOtherRest;
+import ml.comet.experiment.impl.rest.MetricRest;
+import ml.comet.experiment.impl.rest.MinMaxResponse;
+import ml.comet.experiment.impl.rest.OutputUpdate;
+import ml.comet.experiment.impl.rest.ParameterRest;
+import ml.comet.experiment.impl.rest.TagsResponse;
 import ml.comet.experiment.impl.utils.AssetUtils;
 import ml.comet.experiment.impl.utils.JsonUtils;
