--- comet-java-client/src/main/java/ml/comet/experiment/impl/model/GetProjectsResponse.java
@@ -1,18 +1,0 @@
-package ml.comet.experiment.impl.model;
-
-import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.util.List;
-
-@Data
-@AllArgsConstructor
-@NoArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
-@JsonIgnoreProperties(ignoreUnknown = true)
-public class GetProjectsResponse {
-    private List<RestProject> projects;
-}
