+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/AddExperimentTagsRest.java
@@ -1,18 +1,20 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 import java.util.List;
 
 @Data
-@AllArgsConstructor
+@EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
+@AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GetWorkspacesResponse {
-    private List<String> workspaceNames;
+public class AddExperimentTagsRest extends BaseExperimentObject {
+    private List<String> addedTags;
 }
