+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/CreateExperimentResponse.java
@@ -1,18 +1,21 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
-@NoArgsConstructor
+@EqualsAndHashCode(callSuper = true)
 @AllArgsConstructor
+@NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class TagsResponse {
-    private List<String> tags;
+public class CreateExperimentResponse extends BaseExperimentObject {
+    private String workspaceName;
+    private String projectName;
+    private String link;
+    private String name;
 }
