+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/ExperimentStatusResponse.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,14 +6,16 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+/**
+ * Describes response received when experiment set status sent.
+ */
 @Data
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class OutputLine {
-    private String output;
-    private boolean stderr = false;
-    private long localTimestamp;
-    private Long offset;
+public class ExperimentStatusResponse {
+    private long isAliveBeatDurationMillis;
+    private long gpuMonitorIntervalMillis;
+    private long cpuMonitorIntervalMillis;
 }
