+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/ExperimentTimeRequest.java
@@ -1,18 +1,19 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
-@AllArgsConstructor
+@EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
+@AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GetExperimentsResponse {
-    List<ExperimentMetadataRest> experiments;
+public class ExperimentTimeRequest extends BaseExperimentObject {
+    private Long startTimeMillis;
+    private Long endTimeMillis;
 }
