+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/GetGraphResponse.java
@@ -1,22 +1,16 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 @Data
-@EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class ParameterRest extends BaseExperimentObject {
-    private String parameterName;
-    private String parameterValue;
-    private Long step;
-    private Long timestamp;
-    private String context;
+public class GetGraphResponse {
+    private String graph;
 }
