+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/GetHtmlResponse.java
@@ -1,21 +1,16 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
 @Data
-@EqualsAndHashCode(callSuper = true)
 @AllArgsConstructor
 @NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class CreateExperimentResponse extends BaseExperimentObject {
-    private String workspaceName;
-    private String projectName;
-    private String link;
-    private String name;
+public class GetHtmlResponse {
+    private String html;
 }
