+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/GetWorkspacesResponse.java
@@ -1,20 +1,18 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
-@EqualsAndHashCode(callSuper = true)
-@NoArgsConstructor
 @AllArgsConstructor
+@NoArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class LogOtherRest extends BaseExperimentObject {
-    private String key;
-    private String value;
-    private Long timestamp;
+public class GetWorkspacesResponse {
+    private List<String> workspaceNames;
 }
