+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/GitMetadataRest.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -6,6 +6,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
+import ml.comet.experiment.model.GitMetaData;
 
 @Data
 @EqualsAndHashCode(callSuper = true)
@@ -19,4 +20,19 @@ public class GitMetadataRest extends BaseExperimentObject {
     private String branch;
     private String parent;
     private String origin;
+
+    /**
+     * Convert this into the {@link GitMetaData} object of the public API.
+     *
+     * @return the initialized {@link GitMetaData} instance.
+     */
+    public GitMetaData toGitMetaData() {
+        GitMetaData g = new GitMetaData();
+        g.setUser(this.user);
+        g.setRoot(this.root);
+        g.setBranch(this.branch);
+        g.setParent(this.parent);
+        g.setOrigin(this.origin);
+        return g;
+    }
 }
