+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/LogOtherRest.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -7,14 +7,14 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class AddExperimentTagsRest extends BaseExperimentObject {
-    private List<String> addedTags;
+public class LogOtherRest extends BaseExperimentObject {
+    private String key;
+    private String value;
+    private Long timestamp;
 }
