+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/MetricRest.java
@@ -1,0 +1,23 @@
+package ml.comet.experiment.impl.rest;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+
+@Data
+@EqualsAndHashCode(callSuper = true)
+@NoArgsConstructor
+@AllArgsConstructor
+@JsonInclude(JsonInclude.Include.NON_NULL)
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class MetricRest extends BaseExperimentObject {
+    private String metricName;
+    private String metricValue;
+    private String context;
+    private Long step;
+    private Long epoch;
+    private Long timestamp;
+}
