+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/OutputLine.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -7,10 +7,13 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 @Data
-@AllArgsConstructor
 @NoArgsConstructor
+@AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GetHtmlResponse {
-    private String html;
+public class OutputLine {
+    private String output;
+    private boolean stderr = false;
+    private long localTimestamp;
+    private Long offset;
 }
