+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/OutputUpdate.java
@@ -1,16 +1,21 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
+import java.util.List;
+
 @Data
+@EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class GetOutputResponse {
-    private String output;
+public class OutputUpdate extends BaseExperimentObject {
+    private String runContext;
+    private List<OutputLine> outputLines;
 }
