+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/ParameterRest.java
@@ -1,4 +1,4 @@
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonInclude;
@@ -7,15 +7,16 @@ import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
 
-import java.util.List;
-
 @Data
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor
 @AllArgsConstructor
 @JsonInclude(JsonInclude.Include.NON_NULL)
 @JsonIgnoreProperties(ignoreUnknown = true)
-public class OutputUpdate extends BaseExperimentObject {
-    private String runContext;
-    private List<OutputLine> outputLines;
+public class ParameterRest extends BaseExperimentObject {
+    private String parameterName;
+    private String parameterValue;
+    private Long step;
+    private Long timestamp;
+    private String context;
 }
