+++ comet-java-client/src/main/java/ml/comet/experiment/impl/rest/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Contains all classes used to model Comet REST API data transfer objects.
  */
-package ml.comet.experiment.impl.model;
+package ml.comet.experiment.impl.rest;
