+++ comet-java-client/src/main/java/ml/comet/experiment/impl/utils/DataUtils.java
@@ -3,16 +3,16 @@ package ml.comet.experiment.impl.utils;
 import lombok.NonNull;
 import lombok.experimental.UtilityClass;
 import ml.comet.experiment.context.ExperimentContext;
-import ml.comet.experiment.impl.model.AddExperimentTagsRest;
-import ml.comet.experiment.impl.model.AddGraphRest;
-import ml.comet.experiment.impl.model.ExperimentTimeRequest;
-import ml.comet.experiment.impl.model.GitMetadataRest;
-import ml.comet.experiment.impl.model.HtmlRest;
-import ml.comet.experiment.impl.model.LogOtherRest;
-import ml.comet.experiment.impl.model.MetricRest;
-import ml.comet.experiment.impl.model.OutputLine;
-import ml.comet.experiment.impl.model.OutputUpdate;
-import ml.comet.experiment.impl.model.ParameterRest;
+import ml.comet.experiment.impl.rest.AddExperimentTagsRest;
+import ml.comet.experiment.impl.rest.AddGraphRest;
+import ml.comet.experiment.impl.rest.ExperimentTimeRequest;
+import ml.comet.experiment.impl.rest.GitMetadataRest;
+import ml.comet.experiment.impl.rest.HtmlRest;
+import ml.comet.experiment.impl.rest.LogOtherRest;
+import ml.comet.experiment.impl.rest.MetricRest;
+import ml.comet.experiment.impl.rest.OutputLine;
+import ml.comet.experiment.impl.rest.OutputUpdate;
+import ml.comet.experiment.impl.rest.ParameterRest;
 import ml.comet.experiment.model.GitMetaData;
 
 import java.util.Collections;
